storeの内容をmainに移動
